Revert "refactor(ai-help): extract Markdown component"
This reverts commit 4e9a454376542464b2f8773f68b19eef9ed75daa.
